+++ e2e/process-services-cloud/form-field/visibility-condition.e2e.ts
@@ -15,18 +15,14 @@
  * limitations under the License.
  */
 
-import { LoginSSOPage, Widget, AppListCloudPage, IdentityService, GroupIdentityService, ApiService, StringUtil, StartTasksCloudPage, TaskFormCloudComponent } from '@alfresco/adf-testing';
-import { browser, by } from 'protractor';
+import { LoginSSOPage, Widget } from '@alfresco/adf-testing';
+import { browser } from 'protractor';
 
 import { AlfrescoApiCompatibility as AlfrescoApi } from '@alfresco/js-api';
 import { NavigationBarPage } from '../../pages/adf/navigationBarPage';
 import { FormCloudDemoPage } from '../../pages/adf/demo-shell/process-services-cloud/cloudFormDemoPage';
 import { checkboxVisibilityFormJson, multipleCheckboxVisibilityFormJson } from '../../resources/forms/checkbox-visibility-condition';
 import { multipleVisibilityFormJson } from '../../resources/forms/multiple-visibility-conditions';
-import { TasksCloudDemoPage } from '../../pages/adf/demo-shell/process-services/tasksCloudDemoPage';
-import { StartProcessPage } from '../../pages/adf/process-services/startProcessPage';
-import { ProcessCloudDemoPage } from '../../pages/adf/demo-shell/process-services/processCloudDemoPage';
-import { ProcessDetailsCloudDemoPage } from '../../pages/adf/demo-shell/process-services-cloud/processDetailsCloudDemoPage';
 
 describe('Visibility conditions - cloud', () => {
 
@@ -246,147 +242,4 @@ describe('Visibility conditions - cloud', () => {
         await widget.textWidget().setValue(widgets.textThreeId, 'something');
         await widget.textWidget().isWidgetVisible(widgets.textOneId);
     });
-});
-
-describe('Task cloud visibility', async () => {
-
-    const loginSSOPage = new LoginSSOPage();
-    const navigationBarPage = new NavigationBarPage();
-    const appListCloudComponent = new AppListCloudPage();
-    const tasksCloudDemoPage = new TasksCloudDemoPage();
-    const startTask = new StartTasksCloudPage();
-    const taskFormCloudComponent = new TaskFormCloudComponent();
-    const startProcessPage = new StartProcessPage();
-    const processCloudDemoPage = new ProcessCloudDemoPage();
-    const processDetailsCloudDemoPage = new ProcessDetailsCloudDemoPage();
-
-    const simpleApp = browser.params.resources.ACTIVITI_CLOUD_APPS.SIMPLE_APP.name;
-    const standaloneTaskName = StringUtil.generateRandomString(5);
-    const processName = StringUtil.generateRandomString(5);
-    let identityService: IdentityService;
-    let groupIdentityService: GroupIdentityService;
-    let testUser, groupInfo;
-    const apiService = new ApiService(browser.params.config.oauth2.clientId, browser.params.config.bpmHost, browser.params.config.oauth2.host, 'BPM');
-
-    beforeAll(async () => {
-        await apiService.login(browser.params.identityAdmin.email, browser.params.identityAdmin.password);
-        identityService = new IdentityService(apiService);
-        groupIdentityService = new GroupIdentityService(apiService);
-        testUser = await identityService.createIdentityUserWithRole(apiService, [identityService.ROLES.ACTIVITI_USER]);
-        groupInfo = await groupIdentityService.getGroupInfoByGroupName('hr');
-        await identityService.addUserToGroup(testUser.idIdentityService, groupInfo.id);
-
-        await loginSSOPage.loginSSOIdentityService(testUser.email, testUser.password);
-    });
-
-    afterAll(async () => {
-        await apiService.login(browser.params.identityAdmin.email, browser.params.identityAdmin.password);
-        await identityService.deleteIdentityUser(testUser.idIdentityService);
-
-    });
-
-    beforeEach(async () => {
-        await navigationBarPage.navigateToProcessServicesCloudPage();
-        await appListCloudComponent.checkApsContainer();
-        await appListCloudComponent.goToApp(simpleApp);
-        await tasksCloudDemoPage.taskListCloudComponent().checkTaskListIsLoaded();
-    });
-
-    it('[C315170] Should be able to complete a task with a form with required number widgets', async () => {
-        await tasksCloudDemoPage.openNewTaskForm();
-        await startTask.checkFormIsDisplayed();
-        await startTask.addName(standaloneTaskName);
-        await startTask.selectFormDefinition(browser.params.resources.ACTIVITI_CLOUD_APPS.SIMPLE_APP.forms.requirednumbervisibility.name);
-
-        await startTask.clickStartButton();
-        await tasksCloudDemoPage.taskListCloudComponent().selectRow(standaloneTaskName);
-
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Number1');
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Number2');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-
-        await taskFormCloudComponent.formFields().setFieldValue(by.id, 'Number1', 5);
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Number2');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeFalsy();
-
-        await taskFormCloudComponent.formFields().setFieldValue(by.id, 'Number1', 123);
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Number2');
-    });
-
-    it('[C315169] Should be able to start a process with visibility condition for number widgets', async () => {
-
-        await processCloudDemoPage.openNewProcessForm();
-        await startProcessPage.clearField(startProcessPage.processNameInput);
-        await startProcessPage.selectFromProcessDropdown(browser.params.resources.ACTIVITI_CLOUD_APPS.SIMPLE_APP.processes.numbervisibilityprocess);
-
-        await startProcessPage.enterProcessName(processName);
-        await startProcessPage.clickStartProcessButton();
-
-        await processDetailsCloudDemoPage.selectProcessTaskByName(processName);
-        await tasksCloudDemoPage.taskListCloudComponent().selectRow('number_visibility_task');
-        await taskFormCloudComponent.clickClaimButton();
-
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Number1');
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Number2');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-
-        await taskFormCloudComponent.formFields().setFieldValue(by.id, 'Number1', 5);
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Number2');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeFalsy();
-
-        await taskFormCloudComponent.formFields().setFieldValue(by.id, 'Number1', 123);
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Number2');
-
-        await taskFormCloudComponent.formFields().setFieldValue(by.id, 'Number1', 4);
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeFalsy();
-        await taskFormCloudComponent.clickCompleteButton();
-    });
-
-    it('[C315232] Should be able to complete a process with visibility condition for boolean widgets', async () => {
-
-        await processCloudDemoPage.openNewProcessForm();
-        await startProcessPage.clearField(startProcessPage.processNameInput);
-        await startProcessPage.selectFromProcessDropdown(browser.params.resources.ACTIVITI_CLOUD_APPS.SIMPLE_APP.processes.booleanvisibilityprocess);
-        await startProcessPage.enterProcessName(processName);
-        await startProcessPage.clickStartProcessButton();
-
-        await processDetailsCloudDemoPage.selectProcessTaskByName(processName);
-        await tasksCloudDemoPage.taskListCloudComponent().selectRow('boolean_visibility_task');
-        await taskFormCloudComponent.clickClaimButton();
-
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Checkbox2');
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Checkbox3');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-
-        await taskFormCloudComponent.formFields().clickField(by.id, 'Checkbox1');
-        await taskFormCloudComponent.formFields().clickField(by.id, 'Checkbox2');
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Checkbox3');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeFalsy();
-
-        await taskFormCloudComponent.clickCompleteButton();
-    });
-
-    it('[C315208] Should be able to complete a task with Checkbox widgets', async () => {
-        await tasksCloudDemoPage.openNewTaskForm();
-        await startTask.checkFormIsDisplayed();
-        await startTask.addName(standaloneTaskName);
-        await startTask.selectFormDefinition(browser.params.resources.ACTIVITI_CLOUD_APPS.SIMPLE_APP.forms.booleanvisibility.name);
-
-        await startTask.clickStartButton();
-        await tasksCloudDemoPage.taskListCloudComponent().selectRow(standaloneTaskName);
-
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Checkbox2');
-        await taskFormCloudComponent.formFields().checkWidgetIsHidden('Checkbox3');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeTruthy();
-
-        await taskFormCloudComponent.formFields().clickField(by.id, 'Checkbox1');
-        await taskFormCloudComponent.formFields().clickField(by.id, 'Checkbox2');
-        await taskFormCloudComponent.formFields().checkWidgetIsVisible('Checkbox3');
-        await expect(await taskFormCloudComponent.formFields().isCompleteFormButtonDisabled()).toBeFalsy();
-
-        await taskFormCloudComponent.clickCompleteButton();
-    });
-
 });
