+++ e2e/process-services-cloud/task-list-properties.e2e.ts
@@ -366,123 +366,5 @@ describe('Edit task filters and task list properties', () => {
         });
 
     });
-
-    describe('Edit task filters and task list properties - sort properties', () => {
-
-        beforeEach(async () => {
-            await navigationBarPage.navigateToProcessServicesCloudPage();
-            await appListCloudComponent.checkApsContainer();
-            await appListCloudComponent.goToApp(simpleApp);
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().openFilter();
-            await tasksCloudDemoPage.myTasksFilter().checkTaskFilterIsDisplayed();
-
-        });
-
-        it('[C306901] Should display tasks sorted by task name when taskName is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ASSIGNED');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Name');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Name')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Name')).toBe(true);
-
-        });
-
-        it('[C290156] Should display tasks ordered by id when Id is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ASSIGNED');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Id');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Id')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Id')).toBe(true);
-
-        });
-
-        it('[C306903] Should display tasks sorted by processDefinitionId when processDefinitionId is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ASSIGNED');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('ProcessDefinitionId');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'ProcessDefinitionId')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'ProcessDefinitionId')).toBe(true);
-        });
-
-        it('[C306905] Should display tasks sorted by processInstanceId when processInstanceId is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ASSIGNED');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('ProcessInstanceId');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'ProcessInstanceId')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'ProcessInstanceId')).toBe(true);
-        });
-
-        it('[C306907] Should display tasks sorted by assignee when assignee is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().clearAssignee();
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ALL');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Assignee');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Assignee')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Assignee')).toBe(true);
-        });
-
-        it('[C306911] Should display tasks sorted by parentTaskId when parentTaskId is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().clearAssignee();
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ALL');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('ParentTaskId');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'ParentTaskId')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'ParentTaskId')).toBe(true);
-        });
-
-        it('[C290087] Should display tasks ordered by priority when Priority is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().clearAssignee();
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ALL');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Priority');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Priority')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Priority')).toBe(true);
-
-        });
-
-        it('[C307114] Should display tasks sorted by standalone when standalone is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().clearAssignee();
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ALL');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Standalone');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Standalone')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Standalone')).toBe(true);
-        });
-
-        it('[C307115] Should display tasks sorted by owner when owner is selected from sort dropdown', async () => {
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().clearAssignee();
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setStatusFilterDropDown('ALL');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setSortFilterDropDown('Owner');
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('ASC');
-
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('ASC', 'Owner')).toBe(true);
-
-            await tasksCloudDemoPage.editTaskFilterCloudComponent().setOrderFilterDropDown('DESC');
-            await expect(await tasksCloudDemoPage.taskListCloudComponent().getDataTable().checkListIsSorted('DESC', 'Owner')).toBe(true);
-        });
-    });
-
 });
+
